DCS-696: changes following PR comments
Renamed appointmentFactory() to createAppointmentRequest()
Moved createAppointmentRequest() from selectCourtAppointmentRooms.js to appointmentService.js

In whereaboutsApi.js changed method name  from addVideoLinkAppointment to videoLinkBooking and the endpoint called from /court/add-video-link-appointment to video-link-booking. Updated dependent modules.

Updated test files
