+++ backend/api/whereaboutsApi.js
@@ -6,13 +6,13 @@ const whereaboutsApiFactory = client => {
 
   const getCourtLocations = context => get(context, '/court/all-courts')
 
-  const addVideoLinkAppointment = (context, body) => post(context, '/court/add-video-link-appointment', body)
+  const videoLinkBooking = (context, body) => post(context, '/video-link-booking', body)
 
   const getVideoLinkAppointments = (context, body) => post(context, '/court/video-link-appointments', body)
 
   return {
     getCourtLocations,
-    addVideoLinkAppointment,
+    videoLinkBooking,
     getVideoLinkAppointments,
   }
 }
