+++ integration-tests/plugins/index.js
@@ -31,7 +31,7 @@ module.exports = on => {
     stubUser: (username, caseload) => auth.stubUser(username, caseload),
     stubCourts: courts => whereabouts.stubCourtLocations(courts),
     stubGroups: caseload => whereabouts.stubGroups(caseload),
-    stubAddVideoLinkAppointment: appointment => whereabouts.stubAddVideoLinkAppointment(appointment),
+    stubVideoLinkBooking: appointment => whereabouts.stubVideoLinkBooking(appointment),
 
     stubVerifyToken: (active = true) => tokenverification.stubVerifyToken(active),
     stubLoginPage: auth.redirect,
